refactor: input 컴포넌트 컨벤션에 맞게 수정
