+++ src/components/Input/InputStyle.tsx
@@ -1,33 +1,15 @@
 import styled from "styled-components";
-import { FC } from "react";
 
-interface InputProps {
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  value: string;
-  placeholder?: string; 
-}
-
-const InputStyle: FC<InputProps> = ({ onChange, value, placeholder }) => {
-  return (
-    <InputBox>
-      <StyledInput
-        onChange={onChange}
-        value={value}
-        placeholder={placeholder}
-      />
-    </InputBox>
-  );
-};
-
-const InputBox = styled.div`
+export const InputBox = styled.div`
+  width: 100%;
   height: 47px;
   border: 1px solid var(--gray--1);
-  border-radius: 20px;
+  border-radius: 14px;
   display: flex;
   align-items: center;
 `;
 
-const StyledInput = styled.input`
+export const StyledInput = styled.input`
   height: 17px;
   border: none;
   border-radius: 20px;
@@ -36,5 +18,3 @@ const StyledInput = styled.input`
   font-weight: 400;
   padding: 15px 10px;
 `;
-
-export default InputStyle;
